+++ example/lib/main.dart
@@ -1,12 +1,13 @@
 import 'package:env_reader/env_reader.dart';
 import 'package:example/src/env_model.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 Future<void> main(List<String> arguments) async {
   WidgetsFlutterBinding.ensureInitialized();
   await Env.load(
-    source: EnvLoader.asset('assets/env/.env'),
-    password: "MyStrongPassword",
+    EnvLoader.string(await rootBundle.loadString('assets/env/.env')),
+    "MyOptionalSecretKey",
   );
   runApp(
     const MaterialApp(