+++ CHANGELOG.md
@@ -28,3 +28,11 @@
 
 ## 1.6.0
 * Update pubspec
+
+## 1.7.0
+* Change loader format
+* Remove asset loader
+* Update logo
+* Update documentation
+* Add flag obfuscate
+* Remove flutter from pubspec